requirements.txt
gTTS==2.3.2
google-generativeai==0.3.2
opencv-python==4.8.1.78
RPi.GPIO==0.7.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.0
pygame==2.5.2
